edited line 31 & 32

+++ player_delete.php
@@ -28,8 +28,8 @@
     <link rel="stylesheet" href="form.css"> 
     <link rel="stylesheet" href="button.css"> 
      
-    <h3 class="long">DELETE PLAYER</h3> 
-    <form class="long" action="player_delete.php" method="post"> 
+    <h3 class="short">DELETE PLAYER</h3> 
+    <form class="short" action="player_delete.php" method="post"> 
         <table> 
             <tr> 
                 <td>IcNumber</td>
